Extract model updates to a @ModelAttribute method in HomeController
Extract model updates to a single point of contro in @ModelAttribute method addAtributes(). Refactor all other controller methods to redirect to home after executing, and to change the destination of the success/error messages from the model to the redirectAttribute, so they persist after redirection (needed for the execution of addAttributes() at the right time).